% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREannotation-package.R
\docType{package}
\name{CENTREannotation-package}
\alias{CENTREannotation}
\alias{CENTREannotation-package}
\title{CENTREannotation: Hub package for the annotation data of CENTRE
(GENCODE v40 and SCREEN v3)}
\format{
SQLite databases with GENCODE v40 and SCREEN v3 data.
}
\source{
\url{https://www.gencodegenes.org/human/release_40.html}
\url{https://screen.encodeproject.org}
}
\description{
This is an AnnotationHub package for the CENTRE Bioconductor
software package. It contains the GENCODE version 40 annotation and ENCODE
Registry of candidate cis-regulatory elements (cCREs) version 3. All
}
\examples{
\donttest{
library(AnnotationHub)
hub <- AnnotationHub()
eh <- query(hub, "CENTREannotation")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/slrvv/CENTREannotation}
  \item Report bugs at \url{https://github.com/slrvv/CENTREannotation/issues}
}

}
\author{
\strong{Maintainer}: Sara Lopez Ruiz de Vargas \email{lopez_s@molgen.mpg.de} (\href{https://orcid.org/0000-0002-1290-4774}{ORCID})

}
\keyword{internal}
