% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREannotenhDb-obj.R
\docType{data}
\name{CENTREannotenhDb-obj}
\alias{CENTREannotenhDb-obj}
\alias{CENTREannotenhDb}
\title{CENTREannotenhDb: Object interface to SCREEN v3 database}
\format{
An object of class \code{CENTREannotDb} of length 1.
}
\source{
\url{https://screen.encodeproject.org}
}
\usage{
CENTREannotenhDb
}
\description{
Object of class CENTREannotDb that gives acces to the SCREEN v3 database.
}
\examples{
tables(CENTREannotenhDb)
}
\keyword{datasets}
