% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENTREannotgeneDb-obj.R
\docType{data}
\name{CENTREannotgeneDb-obj}
\alias{CENTREannotgeneDb-obj}
\alias{CENTREannotgeneDb}
\title{CENTREannotgeneDb: Object interface to GENCODE v40 database}
\format{
An object of class \code{CENTREannotDb} of length 1.
}
\source{
\url{https://www.gencodegenes.org/human/release_40.html}
}
\usage{
CENTREannotgeneDb
}
\description{
Object of class CENTREannotDb that gives acces to the GENCODE v40 database.
}
\examples{
tables(CENTREannotgeneDb)
}
\keyword{datasets}
